+++ cypress/helpers.ts
@@ -1,0 +1,3 @@
+export function fixed(nr: string | number) {
+  return parseFloat(parseFloat(nr.toString()).toPrecision(2));
+}