+++ cypress/integration/basic.test.ts
@@ -1,6 +1,11 @@
 /// <reference types="Cypress" />
+
+import DeepState from 'deep-state-observer';
+import { GSTCState } from '../../dist/gstc';
+
 describe('Basic', () => {
   it('Simple ESM', () => {
+    let state: DeepState<GSTCState>;
     cy.visit('/examples/simple/simple.esm.html')
       .get('.gstc__chart-timeline-items-row-item[data-gstcid="gstcid-1"] .gstc__chart-timeline-items-row-item-label')
       .should(($el) => {
@@ -8,9 +13,11 @@ describe('Basic', () => {
       })
       .window()
       .then((win) => {
-        expect(win.state.get('config.chart.items.gstcid-1.label')).to.eq('Item 1');
-        win.state.update('config.chart.items.gstcid-1.label', 'New label 1');
-        expect(win.state.get('config.chart.items.gstcid-1.label')).to.eq('New label 1');
+        // @ts-ignore
+        state = win.state as DeepState;
+        expect(state.get('config.chart.items.gstcid-1.label')).to.eq('Item 1');
+        state.update('config.chart.items.gstcid-1.label', 'New label 1');
+        expect(state.get('config.chart.items.gstcid-1.label')).to.eq('New label 1');
       })
       .get('.gstc__chart-timeline-items-row-item[data-gstcid="gstcid-1"] .gstc__chart-timeline-items-row-item-label')
       .should(($el) => {
@@ -19,6 +26,7 @@ describe('Basic', () => {
   });
 
   it('Simple UMD', () => {
+    let state: DeepState<GSTCState>;
     cy.visit('/examples/simple/simple.umd.html')
       .get('.gstc__chart-timeline-items-row-item[data-gstcid="gstcid-1"] .gstc__chart-timeline-items-row-item-label')
       .should(($el) => {
@@ -26,9 +34,11 @@ describe('Basic', () => {
       })
       .window()
       .then((win) => {
-        expect(win.state.get('config.chart.items.gstcid-1.label')).to.eq('Item 1');
-        win.state.update('config.chart.items.gstcid-1.label', 'New label 1');
-        expect(win.state.get('config.chart.items.gstcid-1.label')).to.eq('New label 1');
+        // @ts-ignore
+        state = win.state as DeepState;
+        expect(state.get('config.chart.items.gstcid-1.label')).to.eq('Item 1');
+        state.update('config.chart.items.gstcid-1.label', 'New label 1');
+        expect(state.get('config.chart.items.gstcid-1.label')).to.eq('New label 1');
       })
       .get('.gstc__chart-timeline-items-row-item[data-gstcid="gstcid-1"] .gstc__chart-timeline-items-row-item-label')
       .should(($el) => {