+++ cypress/integration/calendar-dates.test.ts
@@ -1,13 +1,22 @@
 /// <reference types="Cypress" />
+
+import DeepState from 'deep-state-observer';
+import { Api } from '../../dist/api/api';
+import { DataChartTime, GSTCState } from '../../dist/gstc';
+import { fixed } from '../helpers';
+
 describe('Calendar dates', () => {
   it('should generate all dates', () => {
-    let window, merge;
+    let window, merge, state: DeepState<GSTCState>;
     cy.visit('/examples/complex-1/index.html')
       .window()
       .then((win) => {
         window = win;
+        // @ts-ignore
         merge = win.gstc.api.mergeDeep;
-        const time = merge({}, win.state.get('$data.chart.time'));
+        // @ts-ignore
+        state = win.state;
+        const time: DataChartTime = state.get('$data.chart.time');
         expect(time.allDates.length).to.eq(2);
         expect(time.allDates[0].length).to.eq(2);
         cy.log('All dates', time.allDates);
@@ -31,7 +40,7 @@ describe('Calendar dates', () => {
       //
       .get('.gstc__chart-calendar-date--level-1')
       .then(($el) => {
-        const time = merge({}, window.state.get('$data.chart.time'));
+        const time: DataChartTime = state.get('$data.chart.time');
         const first = $el.get(0);
         const day = Cypress.$(first).find('.gstc__chart-calendar-date-content.gstc-date-top').get(0);
         const firstDate = time.levels[1][0];
@@ -41,34 +50,24 @@ describe('Calendar dates', () => {
         expect(time.levels[0].length).to.eq(1);
         cy.log('firstDate', firstDate);
         expect(firstDate.leftGlobalDate.valueOf()).to.eq(time.fromDate.valueOf());
-        const elementWidth = parseFloat(parseFloat(first.style.width).toFixed(4));
-        const dataWidth = parseFloat(firstDate.currentView.width.toFixed(4));
+        const elementWidth = fixed(first.style.width);
+        const dataWidth = fixed(firstDate.currentView.width);
         expect(dataWidth).to.eq(elementWidth);
         expect(dataWidth).to.be.greaterThan(80);
       })
       //
       // move horizontal scroll bar a little bit
       //
-      .get('.gstc__scroll-bar-inner--horizontal')
-      .then(($scrollBar) => {
-        return cy
-          .get('.gstc__scroll-bar-inner--horizontal')
-          .trigger('pointerdown', { screenX: $scrollBar.offset().left + 10, screenY: $scrollBar.offset().top + 10 })
-          .trigger('pointermove', {
-            screenX: $scrollBar.offset().left + 10 + 10,
-            screenY: $scrollBar.offset().top + 10,
-          })
-          .trigger('pointerup');
-      })
+      .scrollH(10)
       .get('.gstc__chart-calendar-date--level-1')
       .then(($el) => {
-        const time = merge({}, window.state.get('$data.chart.time'));
+        const time: DataChartTime = state.get('$data.chart.time');
         const first = $el.get(0);
         const firstDate = time.levels[1][0];
-        const elementWidth = parseFloat(parseFloat(first.style.width).toFixed(4));
-        const dataWidth = parseFloat(firstDate.currentView.width.toFixed(4));
+        const elementWidth = fixed(first.style.width);
+        const dataWidth = fixed(firstDate.currentView.width);
         expect(elementWidth).to.eq(dataWidth);
-        expect(elementWidth).to.be.lessThan(parseFloat(firstDate.width.toFixed(4)));
+        expect(elementWidth).to.be.lessThan(fixed(firstDate.width));
         const day = Cypress.$(first).find('.gstc__chart-calendar-date-content.gstc-date-top').get(0);
         expect(day.textContent).to.eq('01');
       })
@@ -91,10 +90,10 @@ describe('Calendar dates', () => {
         const time = merge({}, window.state.get('$data.chart.time'));
         const last = $el.get($el.length - 1);
         const lastDate = time.levels[1][time.levels[1].length - 1];
-        const elementWidth = parseFloat(parseFloat(last.style.width).toFixed(4));
-        const dataWidth = parseFloat(lastDate.currentView.width.toFixed(4));
+        const elementWidth = fixed(last.style.width);
+        const dataWidth = fixed(lastDate.currentView.width);
         expect(elementWidth).to.eq(dataWidth);
-        expect(elementWidth).to.eq(parseFloat(lastDate.width.toFixed(4)));
+        expect(elementWidth).to.eq(fixed(lastDate.width));
         const day = Cypress.$(last).find('.gstc__chart-calendar-date-content.gstc-date-top').get(0);
         expect(day.textContent).to.eq('31');
       })
@@ -117,10 +116,10 @@ describe('Calendar dates', () => {
         const time = merge({}, window.state.get('$data.chart.time'));
         const last = $el.get($el.length - 1);
         const lastDate = time.levels[1][time.levels[1].length - 1];
-        const elementWidth = parseFloat(parseFloat(last.style.width).toFixed(4));
-        const dataWidth = parseFloat(lastDate.currentView.width.toFixed(4));
+        const elementWidth = fixed(last.style.width);
+        const dataWidth = fixed(lastDate.currentView.width);
         expect(elementWidth).to.eq(dataWidth);
-        expect(elementWidth).to.be.lessThan(parseFloat(lastDate.width.toFixed(4)));
+        expect(elementWidth).to.be.lessThan(fixed(lastDate.width));
         const day = Cypress.$(last).find('.gstc__chart-calendar-date-content.gstc-date-top').get(0);
         expect(day.textContent).to.eq('31');
       });